backend/pyproject.toml
[project]
name = "poc-apl-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "langchain>=1.0.0",
    "langchain-openai",
    "langchain-anthropic",
    "python-dotenv",
    "httpx",
    "asammdf",
    "numpy",
    "psycopg2-binary",
]

[tool.uv]
package = false
